limit sourcemap copy process to deploy

+++ packages/sst/src/runtime/handlers.ts
@@ -140,7 +140,7 @@ export const useRuntimeHandlers = Context.memo(() => {
         if (func.hooks?.afterBuild) await func.hooks.afterBuild(func, out);
 
         let sourcemap: string | undefined;
-        if (built.sourcemap) {
+        if (built.sourcemap && mode === "deploy") {
           const data = await fs.readFile(built.sourcemap);
           await fs.rm(built.sourcemap);
           const hash = crypto.createHash("md5").update(data).digest("hex");
